optimized a lot of things

+++ v6/src/utils/random.hpp
@@ -3,7 +3,7 @@
 class unfiorm {
 private:
 	std::vector<std::default_random_engine> generators;
-	std::vector<std::uniform_real_distribution<float>> distributions;
+	std::vector<std::uniform_real_distribution<double>> distributions;
 
 public:
 	unfiorm() {
@@ -14,13 +14,13 @@ public:
 		num_thread = omp_get_num_threads();
 
 		generators = std::vector<std::default_random_engine>(num_thread);
-		distributions = std::vector<std::uniform_real_distribution<float>>(num_thread);
+		distributions = std::vector<std::uniform_real_distribution<double>>(num_thread);
 
 		for (int i = 0; i < num_thread; ++i)
-			distributions[i] = std::uniform_real_distribution<float>(0, 1);
+			distributions[i] = std::uniform_real_distribution<double>(0, 1);
 	}
 
-	float operator()() {
+	double operator()() {
 		int thread_id = omp_get_thread_num();
 		return distributions[thread_id](generators[thread_id]);
 	}
